refactor on folders and fixed incorrect imports
