+++ plugins/module_utils/pfsense_interface.py
@@ -6,8 +6,8 @@
 from __future__ import absolute_import, division, print_function
 __metaclass__ = type
 import re
-from ansible.module_utils.network.pfsense.pfsense import PFSenseModule, PFSenseModuleBase
-from ansible.module_utils.network.pfsense.pfsense_rule import PFSenseRuleModule
+from ansible_collections.opoplawski.pfsense.plugins.module_utils.pfsense import PFSenseModule, PFSenseModuleBase
+from ansible_collections.opoplawski.pfsense.plugins.module_utils.pfsense_rule import PFSenseRuleModule
 from ansible.module_utils.compat.ipaddress import ip_network
 
 INTERFACE_ARGUMENT_SPEC = dict(
