+++ plugins/module_utils/pfsense_ipsec_p2.py
@@ -5,7 +5,7 @@
 
 from __future__ import absolute_import, division, print_function
 __metaclass__ = type
-from ansible.module_utils.network.pfsense.pfsense import PFSenseModule, PFSenseModuleBase
+from ansible_collections.opoplawski.pfsense.plugins.module_utils.pfsense import PFSenseModule, PFSenseModuleBase
 from copy import deepcopy
 
 IPSEC_P2_ARGUMENT_SPEC = dict(
