+++ plugins/module_utils/pfsense_ipsec_proposal.py
@@ -5,7 +5,7 @@
 
 from __future__ import absolute_import, division, print_function
 __metaclass__ = type
-from ansible.module_utils.network.pfsense.pfsense import PFSenseModule, PFSenseModuleBase
+from ansible_collections.opoplawski.pfsense.plugins.module_utils.pfsense import PFSenseModule, PFSenseModuleBase
 
 
 IPSEC_PROPOSAL_ARGUMENT_SPEC = dict(
