+++ plugins/module_utils/pfsense_rule_separator.py
@@ -6,7 +6,7 @@
 
 from __future__ import absolute_import, division, print_function
 __metaclass__ = type
-from ansible.module_utils.network.pfsense.pfsense import PFSenseModule, PFSenseModuleBase
+from ansible_collections.opoplawski.pfsense.plugins.module_utils.pfsense import PFSenseModule, PFSenseModuleBase
 
 RULE_SEPARATOR_ARGUMENT_SPEC = dict(
     name=dict(required=True, type='str'),
