+++ plugins/modules/pfsense_authserver_ldap.py
@@ -117,7 +117,7 @@ RETURN = """
 """
 
 from ansible.module_utils.basic import AnsibleModule
-from ansible_collections.opoplawski.pfsense.plugins.module_utils.network.pfsense.pfsense import PFSenseModule
+from ansible_collections.opoplawski.pfsense.plugins.module_utils.pfsense import PFSenseModule
 
 
 class PFSenseAuthserverLDAP(object):
