+++ plugins/modules/pfsense_ipsec_proposal.py
@@ -94,7 +94,7 @@ commands:
 """
 
 from ansible.module_utils.basic import AnsibleModule
-from ansible_collections.opoplawski.pfsense.plugins.module_utils.network.pfsense.pfsense_ipsec_proposal import PFSenseIpsecProposalModule, IPSEC_PROPOSAL_ARGUMENT_SPEC, IPSEC_PROPOSAL_REQUIRED_IF
+from ansible_collections.opoplawski.pfsense.plugins.module_utils.pfsense_ipsec_proposal import PFSenseIpsecProposalModule, IPSEC_PROPOSAL_ARGUMENT_SPEC, IPSEC_PROPOSAL_REQUIRED_IF
 
 
 def main():
