+++ plugins/modules/pfsense_user.py
@@ -86,7 +86,7 @@ import base64
 import re
 
 from ansible.module_utils.basic import AnsibleModule
-from ansible_collections.opoplawski.pfsense.plugins.module_utils.network.pfsense.pfsense import PFSenseModule
+from ansible_collections.opoplawski.pfsense.plugins.module_utils.pfsense import PFSenseModule
 
 USER_PHP_COMMAND_PREFIX = """
 require_once('auth.inc');
