+++ plugins/modules/pfsense_vlan.py
@@ -70,7 +70,7 @@ commands:
 """
 
 from ansible.module_utils.basic import AnsibleModule
-from ansible_collections.opoplawski.pfsense.plugins.module_utils.network.pfsense.pfsense_vlan import PFSenseVlanModule, VLAN_ARGUMENT_SPEC
+from ansible_collections.opoplawski.pfsense.plugins.module_utils.pfsense_vlan import PFSenseVlanModule, VLAN_ARGUMENT_SPEC
 
 
 def main():
